Minor refactor and rename
